+++ subversion/libsvn_fs_util/fs-util.c
@@ -2,10 +2,10 @@
  * ends.
  *
  * ====================================================================
- *    Licensed to the Subversion Corporation (SVN Corp.) under one
+ *    Licensed to the Apache Software Foundation (ASF) under one
  *    or more contributor license agreements.  See the NOTICE file
  *    distributed with this work for additional information
- *    regarding copyright ownership.  The SVN Corp. licenses this file
+ *    regarding copyright ownership.  The ASF licenses this file
  *    to you under the Apache License, Version 2.0 (the
  *    "License"); you may not use this file except in compliance
  *    with the License.  You may obtain a copy of the License at
@@ -38,7 +38,7 @@ const char *
 svn_fs__canonicalize_abspath(const char *path, apr_pool_t *pool)
 {
   char *newpath;
-  int path_len;
+  size_t path_len;
   int path_i = 0, newpath_i = 0;
   svn_boolean_t eating_slashes = FALSE;
